generate svgs so they are crispy at lower resolutions
SVG's work well for graphics at higher resolutions but start to break
down at lower resolutions.  For situations like icons, resolutions can get
quite small (i.e. 16x16). As far as I know it's not possible to create a
single SVG that will look good at such a low resolution and at the same
time also look good at lower neighboring resolutions (i.e. 17x17,
18x18, etc).

To address this, I've encoded the Zig logo as a Zig program that will
generates an SVG for any given size. In creating this program I opted
to make things symmetrical. For example, the top/bottom margins now match,
the top/bottom part of the "Z" are now the same height, the gap between the
side components and the center Z is now the same and the angles are more
consistent (topleft angle matches bottomright and bottomleft angle
matches topright). After making things symmetrical the diagonal looked
a bit thicker to me than it was before so I steepend the angle to fix
that.